make current OpenGL context fixed

+++ src/tidop/viewer/group/ModelBase.h
@@ -31,8 +31,8 @@ protected:
 
 	tl::Matrix4x4f modelMatrix;
 
-	unsigned int pointSize;
-	unsigned int lineSize;
+	float pointSize;
+	float lineSize;
 
 public:
